Replacing longs in other places

+++ enclave/core/calls.c
@@ -18,7 +18,7 @@
 #include "report.h"
 #include "td.h"
 
-typedef unsigned long long WORD;
+typedef uint64_t WORD;
 
 #define WORD_SIZE sizeof(WORD)
 
@@ -217,7 +217,7 @@ OE_CATCH:
 **==============================================================================
 */
 
-static void _HandleExit(OE_Code code, long func, uint64_t arg)
+static void _HandleExit(OE_Code code, int64_t func, uint64_t arg)
 {
     OE_Exit(OE_MakeCallArg1(code, func, 0), arg);
 }
@@ -369,7 +369,7 @@ Exit:
 **==============================================================================
 */
 
-static __inline__ void _HandleORET(TD* td, long func, long arg)
+static __inline__ void _HandleORET(TD* td, int64_t func, int64_t arg)
 {
     Callsite* callsite = td->callsites;
 
@@ -480,7 +480,7 @@ OE_Result OE_CallHost(const char* func, void* argsIn)
     }
 
     /* Call into the host */
-    OE_TRY(OE_OCall(OE_FUNC_CALL_HOST, (long)args, NULL, 0));
+    OE_TRY(OE_OCall(OE_FUNC_CALL_HOST, (int64_t)args, NULL, 0));
 
     /* Check the result */
     OE_TRY(args->result);
